Enum autoupdater::parsing::mcp::MaybeCrossPlatform

pub enum MaybeCrossPlatform<T> {
    Cross(T),
    NotCross(HashMap<Platform, T>),
}

This enum allows us to represent a T that may or may not depend on the platform that we wish to install for.

Variants

Cross(T)

This T is the same across all platforms.

NotCross(HashMap<Platform, T>)

This T depends on the Platform

Implementations

impl<T> MaybeCrossPlatform<T>

pub fn get(&self) -> Option<&T>

Gets the T for the current platform.

pub fn get_mut(&mut self) -> Option<&mut T>

pub fn get_owned(self) -> Option<T>